backend/src/utils/jwt.py
from datetime import datetime, timedelta, timezone
from jose import JWTError, jwt
from config import settings

# Example to create a JWT token
def create_jwt_token(email: str):
    expire = datetime.now(timezone.utc) + timedelta(minutes=settings.access_token_expire_minutes)
    to_encode = {
        "sub": email,
        "exp": expire,
        "iat": datetime.now(timezone.utc),
    }
    encoded_jwt = jwt.encode(to_encode, settings.jwt_secret_key, algorithm=settings.jwt_algorithm)
    return encoded_jwt
