backend/pyproject.toml
[project]
name = "nullframe-backend"
version = "0.1.0"
description = "FastAPI backend for Nullframe"
authors = [{ name = "Basel Malalla", email = "malalla.basel@gmail.com" }]
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "motor",
    "python-dotenv",
    "beanie",
    "passlib[bcrypt]",
    "bcrypt==4.0.1",
    "python-jose[cryptography]",
    "pydantic[email]",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["migrations", ".venv"]

select = ["ALL"]
ignore = ["D203", "D213", "ANN001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
